Move page template components to separate files (Gatsby v4 compat change)
