+++ src/cms/preview-templates/AboutPagePreview.js
@@ -1,12 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { AboutPageTemplate } from '../../templates/about-page'
+import { AboutPageTemplate } from '../../templates/AboutPageTemplate'
 
 const AboutPagePreview = ({ entry, widgetFor }) => (
-  <AboutPageTemplate
-    title={entry.getIn(['data', 'title'])}
-    content={widgetFor('body')}
-  />
+  <AboutPageTemplate content={widgetFor('body')} />
 )
 
 AboutPagePreview.propTypes = {
