+++ src/components/all.sass
@@ -773,3 +773,13 @@ a
 @media screen and ( max-width: 768px )
   .wallet-buttons
     display: block
+
+.visuallyhidden
+  border: 0
+  clip: rect(0 0 0 0)
+  height: 1px
+  margin: -1px
+  overflow: hidden
+  padding: 0
+  position: absolute
+  width: 1px
