+++ src/templates/about-page.js
@@ -2,43 +2,17 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { graphql } from 'gatsby'
 import Layout from '../components/Layout'
-import Content, { HTMLContent } from '../components/Content'
-import Logo from '../components/Logo'
+import { HTMLContent } from '../components/Content'
 import BasicHeader from '../components/BasicHeader'
 
-export const AboutPageTemplate = ({ title, content, contentComponent }) => {
-  const PageContent = contentComponent || Content
-
-  return (
-    <section className="section section--gradient section-about-page">
-      <div className="container">
-        <div className="columns">
-          <div className="column is-10 is-offset-1">
-            <div className="section">
-              <h1 className="title is-size-3 has-text-weight-bold is-bold-light has-text-centered" style={{color: 'black'}}>
-                <Logo/>
-              </h1>
-              <PageContent className="content" content={content} />
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
-  )
-}
-
-AboutPageTemplate.propTypes = {
-  title: PropTypes.string.isRequired,
-  content: PropTypes.string,
-  contentComponent: PropTypes.func,
-}
+import { AboutPageTemplate } from './AboutPageTemplate'
 
 const AboutPage = ({ data }) => {
   const { markdownRemark: post } = data
 
   return (
     <Layout>
-      <BasicHeader/>
+      <BasicHeader />
       <AboutPageTemplate
         contentComponent={HTMLContent}
         title={post.frontmatter.title}
